requirements.txt
Django~=3.2.25
djangorestframework~=3.15.1